Fix correctly ignoring undefined docs when importing from rowy
The code intended to do so, but inproperly, it looks like it's
the first time we ran into it. The reason is that when there is a
subtable for a document, deletion in Rowy doesn't trigger deletion
of it, only the top level part.

It would be cool to figure out how to cleanup that garbage in
database, but that's IMHO it's very low priority at the moment.