Fix ignoring undefined firebase docs when importing from rowy (#114)
The code intended to do so, but improperly, it looks like it's the first
time we ran into it. The reason is that when document has a
subcollection, deletion in Rowy doesn't trigger deletion of
subcollection recursively.

It would be cool to figure out how to cleanup that garbage in database,
but that's IMHO it's very low priority at the moment.