Add script and workflow for syncing maps to gdrive (#214)
The script's primary functionality is to make sure that Google Drive
folder in which we store maps as backup has all the live maps, nothing
less nothing more.

There is also a secondary `copy_from` command that is intended for usage
from a one user to copy files owned by some other user. The purpose is
to periodically manually copy the files owned by the service account as
files owned by real user because it's impossible to buy Google Drive
quota for the service account. When then script runs and detects
duplicates, it always prefers to delete the files owned by the service
account.

This resolves #28.